Re-organized the code quite a bit. Took the gazebo ref publisher out of pose optimizer, now its optional, for validation

+++ spinning_lidar_icpslam/src/laser_processing/interrupt_laser_assembler.cpp
@@ -12,7 +12,7 @@ std::string ir_interrupt_topic, assembled_cloud_topic, assemble_service;
 ros::ServiceClient assemble_client;
 laser_assembler::AssembleScans2 assemble_srv;
 ros::Publisher point_cloud_pub;
-int num_points_thesh = 5000;
+size_t num_points_thesh = 5000;
 
 void irInterruptCallback(const std_msgs::Empty::ConstPtr& msg)
 {
